2b finis

+++ scripts/1d-mol.py
@@ -40,7 +40,7 @@ while str(nbuser) != str(rdm):
     elif str(nbuser) is not int :
         print("Ceci n'est pas un nombre, veuillez réessayer")
     
-    elif int(nbuser) < int(rdm):
+    elif float(nbuser) < float(rdm):
         print("C'est plus")
     
     elif int(nbuser) > int(rdm) and int(nbuser) > 0 and int(nbuser) < 100:
